Add files via upload
This Term Project includes the comprehensive marketing and Sales Analysis of Harpic in Pakistani region for 2024. This project is based on overall market of Harpic, Pakistan with regression results of consumer responses on Harpic Consumer purchasing and advertisment behaviour. This research project is useful since it encompasses insights from a Harpic representative about 4 P's of marketing. Moreover, it also includes market visit insights conducted.